+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.0.0: 2023-10-31
+
+* Mobile Safari compatibility
+* Repsonsive fixes
+
 ### 0.9.1: 2023-10-28
 
 * Retain srollability when the posts are deleted after limit is reached